Restructure component and test folders for TDD clarity and maintainability
